+++ s4e-backend/src/main/java/pl/cyfronet/s4e/controller/LoginController.java
@@ -16,7 +16,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import pl.cyfronet.s4e.controller.request.LoginRequest;
 import pl.cyfronet.s4e.controller.response.LoginResponse;
-import pl.cyfronet.s4e.security.JWTTokenService;
+import pl.cyfronet.s4e.security.JwtTokenService;
 
 import javax.validation.Valid;
 
@@ -29,7 +29,7 @@ import static pl.cyfronet.s4e.Constants.API_PREFIX_V1;
 @Tag(name = "login", description = "The Login API")
 public class LoginController {
     private final AuthenticationProvider authenticationProvider;
-    private final JWTTokenService jwtTokenService;
+    private final JwtTokenService jwtTokenService;
 
     @Operation(summary = "Login to the application")
     @ApiResponses({