+++ s4e-backend/src/test/java/pl/cyfronet/s4e/TestJwtUtil.java
@@ -6,19 +6,40 @@ import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.jackson.io.JacksonSerializer;
 import org.springframework.test.web.servlet.request.RequestPostProcessor;
 import pl.cyfronet.s4e.bean.AppUser;
+import pl.cyfronet.s4e.properties.JwtProperties;
 import pl.cyfronet.s4e.security.AppUserDetails;
+import pl.cyfronet.s4e.security.LoadKeyPair;
 
+import java.security.*;
 import java.util.Date;
 import java.util.HashSet;
 import java.util.stream.Collectors;
 
-import static pl.cyfronet.s4e.config.SecurityConfig.JWT_KEY;
-import static pl.cyfronet.s4e.security.JWTTokenService.AUTHORITIES_KEY;
+import static pl.cyfronet.s4e.security.JwtTokenService.AUTHORITIES_KEY;
 
 public class TestJwtUtil {
+    public static final KeyPair JWT_KEY_PAIR;
+    static {
+        try {
+            JwtProperties jwtProperties = new JwtProperties();
+            jwtProperties.setKeyStore("dev_key.p12");
+            jwtProperties.setKeyStorePassword("dev_password");
+            jwtProperties.setKeyAlias("1");
+            jwtProperties.setKeyPassword("dev_password");
+
+            JWT_KEY_PAIR = LoadKeyPair.loadKeyPair(jwtProperties);
+        } catch (Exception e) {
+            throw new IllegalStateException(e);
+        }
+    }
+
     public static RequestPostProcessor jwtBearerToken(AppUser user, ObjectMapper objectMapper) {
+        return jwtBearerToken(user, objectMapper, JWT_KEY_PAIR);
+    }
+
+    public static RequestPostProcessor jwtBearerToken(AppUser user, ObjectMapper objectMapper, KeyPair jwtKeyPair) {
         return mockRequest -> {
-            mockRequest.addHeader("Authorization", "Bearer " + createToken(createAppUserDetails(user), objectMapper));
+            mockRequest.addHeader("Authorization", "Bearer " + createToken(createAppUserDetails(user), objectMapper, jwtKeyPair));
             return mockRequest;
         };
     }
@@ -35,7 +56,7 @@ public class TestJwtUtil {
                 user.isAdmin());
     }
 
-    private static String createToken(AppUserDetails appUserDetails, ObjectMapper objectMapper) {
+    private static String createToken(AppUserDetails appUserDetails, ObjectMapper objectMapper, KeyPair jwtKeyPair) {
         Claims claims = Jwts.claims()
                 .setSubject(appUserDetails.getUsername())
                 .setExpiration(new Date(System.currentTimeMillis() + 10000L));
@@ -44,7 +65,7 @@ public class TestJwtUtil {
                 .collect(Collectors.toList()));
         return Jwts.builder()
                 .setClaims(claims)
-                .signWith(JWT_KEY)
+                .signWith(jwtKeyPair.getPrivate())
                 .serializeToJsonWith(new JacksonSerializer<>(objectMapper))
                 .compact();
     }